move mutt_history.c to history lib
